requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==0.9.2
cryptography==0.8.2
enum34==1.0.4
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.26
MarkupSafe==0.23
msgpack-python==0.4.6
numpy==1.9.2
Pillow==2.8.1
pyasn1==0.1.7
pycparser==2.12
pyOpenSSL==0.15.1
pytz==2015.2
redis==2.10.3
scipy==0.15.1
six==1.9.0
Werkzeug==0.10.4
